+++ src/main/kotlin/com/stack/knowledge/domain/auth/presentation/data/request/GoogleTeacherSignInRequest.kt
@@ -1,5 +1,8 @@
 package com.stack.knowledge.domain.auth.presentation.data.request
 
+import javax.validation.constraints.NotBlank
+
 data class GoogleTeacherSignInRequest(
+    @NotBlank
     val code: String
 )
